--- a/go.mod
+++ b/go.mod
@@ -3,7 +3,7 @@ module github.com/romshark/jscan
 go 1.17
 
 require (
-	github.com/go-faster/jx v0.25.0
+	github.com/go-faster/jx v0.30.0
 	github.com/json-iterator/go v1.1.12
 	github.com/sinhashubham95/jsonic v1.1.0
 	github.com/stretchr/testify v1.7.0
@@ -17,7 +17,9 @@ require (
 	github.com/modern-go/concurrent v0.0.0-20180228061459-e0a39a4cb421 // indirect
 	github.com/modern-go/reflect2 v1.0.2 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/segmentio/asm v1.1.3 // indirect
 	github.com/tidwall/match v1.1.1 // indirect
 	github.com/tidwall/pretty v1.2.0 // indirect
+	golang.org/x/sys v0.0.0-20211110154304-99a53858aa08 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )