```
name                                     old time/op    new time/op    delta
Valid/tiny/gofaster-jx-32                  60.0ns ± 2%    44.1ns ± 2%   -26.57%  (p=0.008 n=5+5)
Valid/small/gofaster-jx-32                  815ns ± 7%     433ns ± 3%   -46.79%  (p=0.008 n=5+5)
Valid/large/gofaster-jx-32                 38.6ms ± 1%    28.3ms ± 2%   -26.56%  (p=0.008 n=5+5)
Valid/unwind_stack/gofaster-jx-32          1.29ms ± 2%    0.71ms ± 5%   -44.38%  (p=0.008 n=5+5)
CalcStats/gofaster-jx/tiny-32              57.6ns ± 3%    51.7ns ± 2%   -10.25%  (p=0.008 n=5+5)
CalcStats/gofaster-jx/small-32              895ns ± 5%     612ns ± 6%   -31.68%  (p=0.008 n=5+5)
CalcStats/gofaster-jx/large-32             40.9ms ± 1%    35.6ms ± 1%   -13.00%  (p=0.008 n=5+5)
CalcStats/gofaster-jx_withpath/tiny-32     69.8ns ± 2%    56.2ns ± 2%   -19.49%  (p=0.008 n=5+5)
CalcStats/gofaster-jx_withpath/small-32    1.53µs ± 2%    0.65µs ± 2%   -57.34%  (p=0.008 n=5+5)
CalcStats/gofaster-jx_withpath/large-32     168ms ± 4%      35ms ± 2%   -79.10%  (p=0.008 n=5+5)

name                                     old alloc/op   new alloc/op   delta
Valid/tiny/gofaster-jx-32                   0.00B          0.00B           ~     (all equal)
Valid/small/gofaster-jx-32                  16.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
Valid/large/gofaster-jx-32                   137B ± 2%       99B ± 7%   -27.34%  (p=0.008 n=5+5)
Valid/unwind_stack/gofaster-jx-32           131kB ± 0%      66kB ± 0%   -49.90%  (p=0.008 n=5+5)
CalcStats/gofaster-jx/tiny-32               0.00B          0.00B           ~     (all equal)
CalcStats/gofaster-jx/small-32              16.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
CalcStats/gofaster-jx/large-32               144B ± 4%      128B ± 3%   -11.11%  (p=0.008 n=5+5)
CalcStats/gofaster-jx_withpath/tiny-32      0.00B          0.00B           ~     (all equal)
CalcStats/gofaster-jx_withpath/small-32     80.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
CalcStats/gofaster-jx_withpath/large-32    52.3MB ± 0%     0.0MB ± 2%  -100.00%  (p=0.008 n=5+5)

name                                     old allocs/op  new allocs/op  delta
Valid/tiny/gofaster-jx-32                    0.00           0.00           ~     (all equal)
Valid/small/gofaster-jx-32                   2.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
Valid/large/gofaster-jx-32                   0.00           0.00           ~     (all equal)
Valid/unwind_stack/gofaster-jx-32           2.05k ± 0%     1.03k ± 0%   -49.90%  (p=0.008 n=5+5)
CalcStats/gofaster-jx/tiny-32                0.00           0.00           ~     (all equal)
CalcStats/gofaster-jx/small-32               2.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
CalcStats/gofaster-jx/large-32               0.00           0.00           ~     (all equal)
CalcStats/gofaster-jx_withpath/tiny-32       0.00           0.00           ~     (all equal)
CalcStats/gofaster-jx_withpath/small-32      13.0 ± 0%       0.0       -100.00%  (p=0.008 n=5+5)
CalcStats/gofaster-jx_withpath/large-32     1.54M ± 0%     0.00M       -100.00%  (p=0.008 n=5+5)
```

Please also update benchmarks in readme when you have time for it :heart: