@ernado 2022-01-15
I've also optimized `withpath` case, reducing allocations